requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.2

Flask-SocketIO==5.3.6
python-socketio==5.11.4
python-engineio==4.10.1
gevent==24.10.1
gevent-websocket==0.10.1